FUCK YOU, const_cast<>()

+++ pliio/tests/test_structcode.py
@@ -6,12 +6,13 @@ from basecase import BaseCase
 #import sys
 #from imread import imread
 from pliio import PyImgC as imgc
+import numpy
 
 class StructCodeTests(BaseCase):
     
     def test_cimage_test_method(self):
         for im in self.imread:
-            imgc.cimage_test(im)
+            imgc.cimage_test(im, dtype=numpy.uint8)
     
     def test_simple_structcodes(self):
         imgc.structcode_parse('B')
